Package List: `booleanGlyph` → `booleanOperations`
This change makes it clearer, perhaps.

I stumbled upon the discrepancy between name and package when I was manually testing the import for all modules.
Since `booleanGlyph` cannot be directly imported, the import threw an error. But since it is part of the `booleanOperations` package, all is good.

(I don’t know why the last line wants to be changed, this is thanks to Github. Feel free to dismiss this PR and do the change manually).